broke it all and reassembled
